Refactor the error handlers
They have now been refactored so that they are per-file. Also note that I renamed the top-level function to resolveErrors so that it does not get confused with handleEndpointMiddleware etc.
